# api/ewbi/swagger.yaml
...
openapi: 3.0.3
info:
  version: 1.0.0
  title: Federation Management Service
  description: |
    # Introduction
    ---
    RESTful APIs that allow an OP to share the edge cloud resources and capabilities securely to other partner OPs over E/WBI.

    ---
    # API Scope

    ---
    APIs defined in this version of the specification can be categorized into the following areas:
    * __FederationManagement__ - Create and manage directed federation relationship with a partner OP
    * __AvailabilityZoneInfoSynchronization__ - Management of resources of partner OP zones and status updates
    * __ArtefactManagement__ - Upload, remove, retrieve and update application descriptors, charts and packages over E/WBI towards a partner OP
    * __FileManagement__ - Upload, remove, retrieve and update application binaries over E/WBI towards a partner OP
    * __ApplicationOnboardingManagement__ - Register, retrieve, update and remove applications over E/WBI towards a partner OP
    * __ApplicationDeploymentManagement__ - Create, update, retrieve and terminate application instances over E/WBI towards a partner OP
    * __AppProviderResourceManagement__ - Static resource reservation for an application provider over E/WBI for partner OP zones
    * __EdgeNodeSharing__ - Edge discovery procedures towards partner OP over E/WBI.
    * __LBORoamingAuthentication__ - Validation of user client authentication from home OP

    ---
    # Definitions
    ---
...

...
    * __OP__ - Operator platform.
...

...

...

...

...

...

...
components:
  schemas:
    InterfaceId:
      type: string
      pattern: ^[a-z0-9]([-a-z0-9]{0,62}[a-z0-9])?$
      description: Each Port and corresponding traffic protocol exposed by the component is identified by a name. Application client on user device requires this to uniquely identify the interface.
    AccessPointInfo:
      description: Information about the IP and Port exposed by the OP. Application clients shall use these access points to reach this application instance.
      type: array
      items:
        type: object
        required:
          - interfaceId
          - accessPoints
        properties:
          interfaceId:
            $ref: "#/components/schemas/InterfaceId"
          accessPoints:
            $ref: "#/components/schemas/ServiceEndpoint"
      minItems: 1
    AppIdentifier:
      type: string
      pattern: ^[A-Za-z][A-Za-z0-9_]{7,63}$
      description: Identifier used to refer to an application.
    AppProviderId:
      type: string
      pattern: ^[A-Za-z][A-Za-z0-9_]{7,63}$
      description: UserId of the app provider.  Identifier is relevant only in context of this federation.
    ArtefactId:
      type: string
      format: uuid
      description: A globally unique identifier associated with the artefact. Originating OP generates this identifier when artefact is submitted over NBI.
...
    InstanceIdentifier:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9_]{6,62}[A-Za-z0-9]$
      description: Unique identifier generated by the partner OP to identify an instance of the application on a specific zone.
    InstanceState:
      type: string
      enum:
        - PENDING
        - READY
        - FAILED
        - TERMINATING
      description: Running status of the application instance.
...
    Ipv4Addr:
      type: string
      pattern: ^(([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])\.){3}([0-9]|[1-9][0-9]|1[0-9][0-9]|2[0-4][0-9]|25[0-5])$
      example: 198.51.100.1
    Ipv6Addr:
      type: string
      allOf:
        - pattern: ^((:|(0?|([1-9a-f][0-9a-f]{0,3}))):)((0?|([1-9a-f][0-9a-f]{0,3})):){0,6}(:|(0?|([1-9a-f][0-9a-f]{0,3})))$
        - pattern: ^((([^:]+:){7}([^:]+))|((([^:]+:)*[^:]+)?::(([^:]+:)*[^:]+)?))$
      example: 2001:db8:85a3::8a2e:370:7334
...
    FixedNetworkIds:
      type: array
      items:
        type: string
      description: List of network identifier associated with the fixed line network of the operator platform.
      minItems: 1
    FederationCallbackId:
      type: string
      pattern: ^[a-z0-9]([-a-z0-9]{0,62}[a-z0-9])?$
      readOnly: true
      description: UUID for grouping resources in the origin OP before a federation context ID exists.
    FederationContextId:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9-]*$
      readOnly: true
      description: This identifier shall be provided by the partner OP on successful verification and validation of the federation create request and is used by partner op to identify this newly created federation context. Originating OP shall provide this identifier in any subsequent request towards the partner op.
...
    FileId:
      type: string
      format: uuid
      description: A globally unique identifier associated with the image file. Originating OP generates this identifier when file is uploaded over NBI.
...
    FlavourId:
      type: string
      description: An identifier to refer to a specific combination of compute resources
    GeoLocation:
      type: string
      description: Latitude,Longitude as decimal fraction up to 4 digit precision
      pattern: ^([-+]?)([\d]{1,2})((((\.)([\d]{1,4}))?(,)))(([-+]?)([\d]{1,3})((\.)([\d]{1,4}))?)$
    Mcc:
      type: string
      pattern: ^\d{3}$
    Mnc:
      type: string
      pattern: ^\d{2,3}$
...
    PoolId:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9_]{6,30}[A-Za-z0-9]$
      description: OP defined Identifier for the pool reserved for the ISV. It should be unique with an OP.
    Port:
      type: integer
      minimum: 0
    Status:
      type: string
      enum:
        - FAILED
        - TEMPORARY_FAILURE
        - AVAILABLE
        - LOCKED
        - NOT_AVAILABLE
...
    Vcpu:
      type: string
      pattern: ^\d+((\.\d{1,3})|(m))?$
      description: Number of vcpus in whole, decimal up to millivcpu, or millivcpu format.
      example:
        whole:
          value: 2
        decimal:
          value: 0.500
        millivcpu:
          value: 500m
...
    ZoneIdentifier:
      type: string
      pattern: ^[A-Za-z0-9][A-Za-z0-9-]*$
      description: Human readable name of the zone.
...
    EdgeAppFQDN:
      type: string
      description: DNS FQDN assigned to application instances in an availability zone. User Clients can resolve the FQDN to communicate with the edge instances of the application
...
    ComputeResourceInfo:
      type: object
      required:
        - cpuArchType
        - numCPU
        - memory
      properties:
        cpuArchType:
          type: string
          enum:
            - ISA_X86_64
            - ISA_ARM_64
          description: CPU Instruction Set Architecture (ISA) E.g., Intel, Arm etc.
        numCPU:
          $ref: '#/components/schemas/Vcpu'
        memory:
          type: integer
          format: int64
          description: Amount of RAM in Mbytes
        diskStorage:
          type: integer
          format: int32
          description: Amount of disk storage in Gbytes for a given ISA type
        gpu:
          type: array
          items:
            $ref: '#/components/schemas/GpuInfo'
        vpu:
          type: integer
          description: Number of Intel VPUs available for a given ISA type
        fpga:
          type: integer
          description: Number of FPGAs available for a given ISA type
        hugepages:
          type: array
          items:
            $ref: '#/components/schemas/HugePage'
        cpuExclusivity:
          type: boolean
          description: Support for exclusive CPUs
...
    GpuInfo:
      type: object
      required:
        - gpuVendorType
        - gpuModeName
        - gpuMemory
        - numGPU
      properties:
        gpuVendorType:
          type: string
          enum:
            - GPU_PROVIDER_NVIDIA
            - GPU_PROVIDER_AMD
          description: GPU vendor name e.g. NVIDIA, AMD etc.
          example: Nvidia
        gpuModeName:
          type: string
          description: Model name corresponding to vendorType may include info e.g. for NVIDIA, model name could be “Tesla M60”, “Tesla V100” etc.
        gpuMemory:
          type: integer
          description: GPU memory in Mbytes
        numGPU:
          type: integer
          description: Number of GPUs
    HugePage:
      type: object
      required:
        - pageSize
        - number
      properties:
        pageSize:
          type: string
          enum:
            - 2MB
            - 4MB
            - 1GB
          description: Size of hugepage
        number:
          type: integer
          description: Total number of huge pages
...
    InvalidParam:
      type: object
      properties:
        param:
          type: string
        reason:
          type: string
      required:
        - param
    MobileNetworkIds:
      type: object
      properties:
        mcc:
          $ref: '#/components/schemas/Mcc'
        mncs:
          type: array
          items:
            $ref: '#/components/schemas/Mnc'
          minItems: 1
...
    ProblemDetails:
      type: object
      properties:
        title:
          type: string
          description: Summary of the problem
        detail:
          type: string
          description: Specific detail of the issue
        cause:
          type: string
          description: Fixed string indicating cause of the issue
        invalidParams:
          type: array
          items:
            $ref: '#/components/schemas/InvalidParam'
          minItems: 0
...
    ServiceEndpoint:
      type: object
      required:
        - port
      anyOf:
        - required:
            - fqdn
        - required:
            - ipv4Addresses
        - required:
            - ipv6Addresses
      properties:
        port:
          $ref: '#/components/schemas/Port'
        fqdn:
          $ref: '#/components/schemas/EdgeAppFQDN'
        ipv4Addresses:
          type: array
          items:
            $ref: '#/components/schemas/Ipv4Addr'
          minItems: 1
        ipv6Addresses:
          type: array
          items:
            $ref: '#/components/schemas/Ipv6Addr'
          minItems: 1
    ZoneDetails:
      type: object
      required:
        - zoneId
        - geographyDetails
      properties:
        zoneId:
          $ref: '#/components/schemas/ZoneIdentifier'
        geolocation:
          $ref: '#/components/schemas/GeoLocation'
        geographyDetails:
          type: string
          description: Details about cities or state covered by the edge. Details about the type of locality for eg rural, urban, industrial etc. This information is defined in human readable form.
...
  responses:
    "400":
      description: Bad request
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    "401":
      description: Unauthorized
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    "404":
      description: Not Found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    "409":
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    "422":
      description: Unprocessable Entity
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    "500":
      description: Internal Server Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    "503":
      description: Service Unavailable
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
    "520":
      description: Web Server Returned an Unknown Error
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetails'
...
    default:
      description: Generic Error
paths:
...
  /{federationCallbackId}/partnerStatusLink:
    post:
      operationId: PartnerStatusLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: |
          OP uses this callback api to notify partner OP about change in federation status, federation metadata or offered zone details. Allowed combinations of objectType and operationType are
          - FEDERATION - STATUS: Status specified by parameter 'federationStatus'.
          - ZONES - STATUS: Status specified by parameter 'zoneStatus'.
          - ZONES - ADD: Use parameter 'addZones' to define add new zones
          - ZONES - REMOVE: Use parameter 'removeZones' to define remove zones.
          - EDGE_DISCOVERY_SERVICE - UPDATE: Use parameter 'edgeDiscoverySvcEndPoint' to specify new endpoints
          - LCM_SERVICE - UPDATE: Use parameter 'lcmSvcEndPoint' to specify new endpoints
          - MOBILE_NETWORK_CODES - ADD: Use parameter 'addMobileNetworkIds' to define new mobile network codes.
          - MOBILE_NETWORK_CODES - REMOVE: Use parameter 'removeMobileNetworkIds' to remove mobile network codes.
          - FIXED_NETWORK_CODES - ADD: Use parameter 'addFixedNetworkIds' to define new fixed network codes.
          - FIXED_NETWORK_CODES - REMOVE: Use parameter 'removeFixedNetworkIds' to remove fixed network codes.
        content:
          application/json:
            schema:
              type: object
              required:
                - federationContextId
                - objectType
                - operationType
                - modificationDate
              properties:
                federationContextId:
                  $ref: "#/components/schemas/FederationContextId"
                objectType:
                  type: string
                  enum:
                    - FEDERATION
                    - ZONES
                    - EDGE_DISCOVERY_SERVICE
                    - LCM_SERVICE
                    - MOBILE_NETWORK_CODES
                    - FIXED_NETWORK_CODES
                operationType:
                  type: string
                  enum:
                    - STATUS
                    - UPDATE
                    - ADD
                    - REMOVE
                edgeDiscoverySvcEndPoint:
                  $ref: "#/components/schemas/ServiceEndpoint"
                lcmSvcEndPoint:
                  $ref: "#/components/schemas/ServiceEndpoint"
                addMobileNetworkIds:
                  $ref: "#/components/schemas/MobileNetworkIds"
                removeMobileNetworkIds:
                  $ref: "#/components/schemas/MobileNetworkIds"
                addFixedNetworkIds:
                  $ref: "#/components/schemas/FixedNetworkIds"
                removeFixedNetworkIds:
                  $ref: "#/components/schemas/FixedNetworkIds"
                addZones:
                  type: array
                  items:
                    $ref: "#/components/schemas/ZoneDetails"
                  description: List of zones, which the operator platform wishes to make available to developers/ISVs of requesting operator platform.
                  minItems: 1
                removeZones:
                  type: array
                  items:
                    $ref: "#/components/schemas/ZoneIdentifier"
                  description: List of zones, which the operator platform no longer wishes to share.
                  minItems: 1
                zoneStatus:
                  type: array
                  items:
                    type: object
                    required:
                      - zoneId
                      - status
                    properties:
                      zoneId:
                        $ref: "#/components/schemas/ZoneIdentifier"
                      status:
                        $ref: "#/components/schemas/Status"
                  minItems: 1
                federationStatus:
                  $ref: "#/components/schemas/Status"
                modificationDate:
                  type: string
                  format: date-time
                  description: Date and time of the federation modification by the originating partner OP
      responses:
        "204":
          description: Expected response to a successful call back processing
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/availZoneNotifLink:
    post:
      operationId: AvailZoneNotifLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification about resource availability.
        content:
          application/json:
            schema:
              type: object
              required:
                - federationContextId
                - zoneId
                - zoneResUpdInfo
              properties:
                federationContextId:
                  $ref: "#/components/schemas/FederationContextId"
                zoneId:
                  $ref: "#/components/schemas/ZoneIdentifier"
                zoneResUpdInfo:
                  type: array
                  items:
                    type: object
                    minProperties: 1
                    properties:
                      availableCompResources:
                        description: Resources exclusively reserved for the originator OP.
                        type: array
                        items:
                          $ref: "#/components/schemas/ComputeResourceInfo"
                        minItems: 1
                      availableNetResources:
                        type: object
                        properties:
                          egressBandWidth:
                            type: integer
                            format: int32
                            description: Max dl throughput that this edge can offer. It is defined in Mbps.
                          dedicatedNIC:
                            type: integer
                            format: int32
                          supportSriov:
                            type: boolean
                            description: If this zone support SRIOV networks or not
                          supportDPDK:
                            type: boolean
                            description: If this zone supports DPDK based networking
                        minProperties: 1
      responses:
        "200":
          description: Zone info notification acknowledged
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/appStatusCallbackLink:
    post:
      operationId: AppStatusCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - appId
                - statusInfo
              properties:
                appId:
                  $ref: "#/components/schemas/AppIdentifier"
                statusInfo:
                  type: array
                  items:
                    type: object
                    required:
                      - zoneId
                      - onboardStatusInfo
                    properties:
                      zoneId:
                        $ref: "#/components/schemas/ZoneIdentifier"
                      onboardStatusInfo:
                        description: Defines change in application status. This change could be related to application itself or an application instance status
                        type: string
                        enum:
                          - PENDING
                          - ONBOARDED
                          - DEBOARDING
                          - REMOVED
                          - FAILED
                  minItems: 1
      responses:
        "204":
          description: Application status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/artefactStatusCallbackLink:
    post:
      operationId: ArtefactStatusCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - artefactId
                - updateStatus
              properties:
                artefactId:
                  $ref: "#/components/schemas/ArtefactId"
                updateStatus:
                  description: Defines change in artefact status. This change could be related to the artefact itself or an artefact instance status
                  type: string
                  enum:
                    - PENDING
                    - ERROR
                    - READY
                    - UNKNOWN
      responses:
        "204":
          description: Artefact status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/fileStatusCallbackLink:
    post:
      operationId: FileStatusCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - fileId
                - updateStatus
              properties:
                fileId:
                  $ref: "#/components/schemas/FileId"
                updateStatus:
                  description: Defines change in file status. This change could be related to the file itself or a file instance status
                  type: string
                  enum:
                    - PENDING
                    - ERROR
                    - READY
                    - UNKNOWN
      responses:
        "204":
          description: File status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/appInstCallbackLink:
    post:
      operationId: AppInstCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - appId
                - appInstanceId
                - zoneId
                - appInstanceInfo
              properties:
                appId:
                  $ref: "#/components/schemas/AppIdentifier"
                appInstanceId:
                  $ref: "#/components/schemas/InstanceIdentifier"
                zoneId:
                  $ref: "#/components/schemas/ZoneIdentifier"
                appInstanceInfo:
                  type: object
                  properties:
                    appInstanceState:
                      $ref: "#/components/schemas/InstanceState"
                    message:
                      type: string
                      description: Event information or failure message.
                    accesspointInfo:
                      $ref: "#/components/schemas/AccessPointInfo"
                  minProperties: 1
                modificationDate:
                  type: string
                  format: date-time
                  description: Date and time of the instance state modification by partner OP.
      responses:
        "204":
          description: Application instance state notification acknowledged
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"
  /{federationCallbackId}/resourceReservationCallbackLink:
    post:
      operationId: ResourceReservationCallbackLink
      parameters:
        - name: federationCallbackId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/FederationCallbackId"
      requestBody:
        description: Notification payload.
        content:
          application/json:
            schema:
              type: object
              required:
                - federationContextId
                - zoneId
                - appProviderId
                - poolId
                - grantedFlavours
              properties:
                federationContextId:
                  $ref: "#/components/schemas/FederationContextId"
                zoneId:
                  $ref: "#/components/schemas/ZoneIdentifier"
                appProviderId:
                  $ref: "#/components/schemas/AppProviderId"
                poolId:
                  $ref: "#/components/schemas/PoolId"
                grantedFlavours:
                  type: array
                  items:
                    type: object
                    required:
                      - flavourId
                      - numFlavour
                    properties:
                      flavourId:
                        $ref: "#/components/schemas/FlavourId"
                      numFlavour:
                        type: integer
                        format: int32
                        description: Count of flavour
                  minItems: 1
      responses:
        "204":
          description: Updated Resource reservation status updated
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "404":
          $ref: "#/components/responses/404"
        "409":
          $ref: "#/components/responses/409"
        "422":
          $ref: "#/components/responses/422"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
        "520":
          $ref: "#/components/responses/520"
        default:
          $ref: "#/components/responses/default"